Split tests into separate files to make them more readable
