+++ tests/test_run_tests_using_dtox.py
@@ -1,0 +1,6 @@
+from run_dtox import run
+
+
+class TestRunDtoxTests:
+
+    pass
